requirements.txt
anyio==4.3.0
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==42.0.7
daphne==4.1.2
Django==5.0.1
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
google-auth==2.29.0
h11==0.14.0
httptools==0.6.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
inflection==0.5.1
packaging==23.2
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.32.2
rsa==4.9
service-identity==24.1.0
setuptools==69.0.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
whitenoise==6.6.0
zope.interface==6.4
